Split `AsyncInitializer` into initializer and server & get rid of `IAsyncDisposable`
Having it separate makes testing easier (you don't have to rely on the
server to be started, e.g. when using `TestServer`).
Using `IAsyncDisposable` poses a problem - containers might know that
pattern and dispose them on their own, which is a problem because that
is not expected at all. It not only might violate the order, it might
also disable some use-cases (i.e. using `IServiceProvider` inside the
Dispose method).